+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/DiscoveryConfigModifier.java
@@ -55,6 +55,8 @@ public class DiscoveryConfigModifier implements PolarisConfigModifier {
 		configuration.getConsumer().getZeroProtection().setEnable(polarisDiscoveryProperties.isZeroProtectionEnabled());
 		configuration.getConsumer().getZeroProtection()
 				.setNeedTestConnectivity(polarisDiscoveryProperties.isZeroProtectionNeedTestConnectivity());
+
+		configuration.getConsumer().getOutlierDetection().setWhen(polarisDiscoveryProperties.getDetect().getWhen());
 	}
 
 	@Override