+++ spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/PolarisDiscoveryPropertiesTest.java
@@ -18,6 +18,7 @@
 package com.tencent.cloud.polaris;
 
 
+import com.tencent.polaris.api.config.consumer.OutlierDetectionConfig;
 import org.junit.jupiter.api.Test;
 
 import static com.tencent.polaris.test.common.Consts.NAMESPACE_TEST;
@@ -97,6 +98,11 @@ public class PolarisDiscoveryPropertiesTest {
 		polarisDiscoveryProperties.setNamespaceExports("*");
 		assertThat(polarisDiscoveryProperties.getNamespaceExports()).isEqualTo("*");
 
+		PolarisDiscoveryProperties.Detect detect = new PolarisDiscoveryProperties.Detect();
+		detect.setWhen(OutlierDetectionConfig.When.after_call);
+		polarisDiscoveryProperties.setDetect(detect);
+		assertThat(polarisDiscoveryProperties.getDetect().getWhen()).isEqualTo(OutlierDetectionConfig.When.after_call);
+
 		assertThat(polarisDiscoveryProperties.toString())
 				.isEqualTo("PolarisDiscoveryProperties{"
 						+ "namespace='Test'"
@@ -114,6 +120,7 @@ public class PolarisDiscoveryPropertiesTest {
 						+ ", zeroProtectionNeedTestConnectivity=false"
 						+ ", preferIpv6=true"
 						+ ", allRecoverEnabled=false"
-						+ ", namespaceExports='*'}");
+						+ ", namespaceExports='*'"
+						+ ", detect=Detect{when=after_call}}");
 	}
 }